requirements.txt
langchain
langchain-openai
langchain-community
openai
requests
python-dotenv
pydantic
streamlit
langchain-google-community
amadeus
langgraph-checkpoint-sqlite
openrouteservice